refactor(test): Organize fixture architecture
