+++ hardhat.config.ts
@@ -236,7 +236,7 @@ const config: HardhatUserConfig = {
   typechain: {
     // outDir: 'src/types', // defaults to './typechain-types/'
     target: 'ethers-v5',
-    // externalArtifacts: [], // optional array of glob patterns with external artifacts to process (for example external libs from node_modules)
+    externalArtifacts: ['./contracts/extensions/bills/lib/**.json'], // optional array of glob patterns with external artifacts to process (for example external libs from node_modules)
     alwaysGenerateOverloads: false, // should overloads with full signatures like deposit(uint256) be generated always, even if there are no overloads?
     dontOverrideCompile: false, // defaults to false
   },
