+++ test/extensions/zap-pools.spec.ts
@@ -3,7 +3,7 @@ import { dex, utils, farmV2 } from '@ape.swap/hardhat-test-helpers'
 import { mine, time, loadFixture } from '@nomicfoundation/hardhat-network-helpers'
 import '@nomicfoundation/hardhat-chai-matchers'
 import { ethers } from 'hardhat'
-import { deployDexAndZap } from '../fixtures/deployDexAndZap'
+import { deployDexAndZap } from '../fixtures'
 
 const ether = utils.ether
