+++ test/fixtures/index.ts
@@ -1,0 +1,3 @@
+export * from './deployBills'
+export * from './deployDexAndZap'
+export * from './deployMockFarmV2'
